Extracted and organized some layout and form components
