+++ src/application/frontend/src/app/(dashboard)/dashboard/layout.tsx
@@ -1,5 +1,4 @@
 'use client';
-
 import { useState } from 'react';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
