+++ src/application/frontend/src/app/(dashboard)/dashboard/page.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { Button } from '@/components/ui/button';
 import {
   Card,
