+++ src/application/frontend/src/app/layout.tsx
@@ -1,5 +1,5 @@
 import './globals.css';
-import { UserProvider } from '@/components/auth/user-context';
+import { UserProvider } from '@/components/features/auth/user-context';
 import type { Metadata, Viewport } from 'next';
 import { Manrope } from 'next/font/google';
 
@@ -26,7 +26,7 @@ export default function RootLayout({
       className={`bg-white dark:bg-gray-950 text-black dark:text-white ${manrope.className}`}
     >
       <body className="min-h-[100dvh] bg-gray-50">
-        {<UserProvider>{children}</UserProvider>}
+        <UserProvider>{children}</UserProvider>
       </body>
     </html>
   );
