+++ src/application/frontend/src/app/page.tsx
@@ -1,17 +1,9 @@
-'use client';
 import Link from 'next/link';
-import {
-  ArrowRight,
-  ShieldCheck,
-  Activity,
-  Sparkles,
-  Brain,
-} from 'lucide-react';
-import { Avatar, AvatarFallback } from '@/components/ui/avatar';
+import { ArrowRight, ShieldCheck, Activity, Sparkles } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Alert } from '@/components/ui/alert';
 import { DUMMY_ALERTS } from '@/utils/testing';
-import { useUser } from '@/hooks/useUser';
+import { Header } from '@/components/layout/header';
 
 const featureCards = [
   {
@@ -35,39 +27,9 @@ const featureCards = [
 ];
 
 export default function HomePage() {
-  const { user } = useUser();
-
   return (
     <main className="bg-gradient-to-b from-orange-50 via-white to-white">
-      <header className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8 py-6 flex items-center justify-between">
-        <Link href="/" className="flex items-center space-x-2">
-          <Brain className="h-9 w-9 text-orange-500" />
-          <span className="text-lg font-semibold text-gray-900">
-            AI Doctors
-          </span>
-        </Link>
-        {user ? (
-          <div className="flex items-center gap-3">
-            <Button asChild className="rounded-full px-4">
-              <Link href="/dashboard">Go to dashboard</Link>
-            </Button>
-            <div className="flex items-center gap-3 rounded-full bg-white border border-gray-200 px-3 py-1.5 shadow-sm">
-              <div className="text-sm text-gray-800">
-                {user.username || 'Account'}
-              </div>
-            </div>
-          </div>
-        ) : (
-          <div className="flex items-center gap-3">
-            <Button asChild variant="ghost" className="rounded-full px-4">
-              <Link href="/login">Log in</Link>
-            </Button>
-            <Button asChild className="rounded-full px-4">
-              <Link href="/signup">Sign up</Link>
-            </Button>
-          </div>
-        )}
-      </header>
+      <Header />
       <section className="relative overflow-hidden">
         <div className="absolute inset-0 pointer-events-none bg-[radial-gradient(circle_at_20%_20%,rgba(255,159,64,0.12),transparent_35%),radial-gradient(circle_at_80%_0%,rgba(59,130,246,0.08),transparent_28%)]" />
         <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8 py-12 relative">
