+++ src/application/frontend/src/components/ui/avatar.tsx
@@ -1,8 +1,5 @@
-'use client';
-
 import * as React from 'react';
 import { Avatar as AvatarPrimitive } from 'radix-ui';
-
 import { cn } from '@/utils/general';
 
 function Avatar({
