+++ src/application/frontend/src/components/ui/button.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { Slot as SlotPrimitive } from 'radix-ui';
 import { cva, type VariantProps } from 'class-variance-authority';
-
 import { cn } from '@/utils/general';
 
 const buttonVariants = cva(
