+++ src/application/frontend/src/components/ui/card.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-
 import { cn } from '@/utils/general';
 
 function Card({ className, ...props }: React.ComponentProps<'div'>) {
