+++ src/application/frontend/src/components/ui/dropdown-menu.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import * as React from 'react';
 import { DropdownMenu as DropdownMenuPrimitive } from 'radix-ui';
 import { CheckIcon, ChevronRightIcon, CircleIcon } from 'lucide-react';
