+++ src/application/frontend/src/components/ui/input.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-
 import { cn } from '@/utils/general';
 
 function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
