+++ src/application/frontend/src/components/ui/radio-group.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import * as React from 'react';
 import { RadioGroup as RadioGroupPrimitive } from 'radix-ui';
 import { CircleIcon } from 'lucide-react';
