+++ src/application/frontend/src/components/ui/textarea.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-
 import { cn } from '@/utils/general';
 
 export interface TextareaProps
