+++ src/application/frontend/src/hooks/useUser.ts
@@ -1,4 +1,5 @@
-import { UserCtx } from '@/components/auth/user-context';
+'use client';
+import { UserCtx } from '@/components/features/auth/user-context';
 import { useContext } from 'react';
 
 export function useUser() {
