--- src/application/frontend/src/types/login-types.ts
@@ -1,12 +1,0 @@
-export type LoginFormValues = {
-  username: string;
-  password: string;
-};
-
-export type SignUpFormValues = {
-  firstName: string;
-  lastName: string;
-  username: string;
-  email: string;
-  password: string;
-};
