+++ src/application/frontend/src/types/user-types.ts
@@ -6,9 +6,3 @@ export type User = {
   disabled: boolean;
   roles: string[];
 };
-
-export type UserContext = {
-  user: User | null;
-  setUser: (u: User | null) => void;
-  isLoading: boolean;
-};
